Added tab map to pokemon.cs
