+++ PokemonGo.RocketAPI.Console/IntegerInput.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Created by SharpDevelop.
- * User: DCRax
+ * User: Xelwon
  * Date: 11/09/2016
  * Time: 3:06
  *
