+++ PokemonGo.RocketAPI.Console/Items.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Created by SharpDevelop.
- * User: DCRax
+ * User: Xelwon
  * Date: 10/09/2016
  * Time: 18:19
  *
