+++ PokemonGo.RocketAPI.Console/Prompt.cs
@@ -1,6 +1,6 @@
 ﻿/*
  * Created by SharpDevelop.
- * User: DCRax
+ * User: Xelwon
  * Date: 13/09/2016
  * Time: 22:37
  *
